[Feature] Color channel filtering. Refactoring
In color channel filtering feature.
Used color enums from Qt namespace (Qt::red, Qt::green, Qt::blue).

+++ imageprocessing.h
@@ -9,7 +9,7 @@ void changeBrigthness(QImage &image, int delta);
 void sharpen(QImage &image);
 void blur(QImage &image);
 void rotate(QImage &image, int angle);
-void filterOutColor(QImage &image, short color);
+void filterOutColor(QImage &image, int color);
 
 #endif // IMAGEPROCESSING
